% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_feed}
\alias{get_feed}
\title{Retrieve a JSON Feed}
\usage{
get_feed(url)
}
\arguments{
\item{url}{the URL of a feed

The key \code{items} contains the information about individual posts.
The top level keys relate to data about the feed itself. If only the meta
data are required, use \link{get_feed_info}, but if only the \code{items}
are required, use \link{get_feed_items}.}
}
\value{
a list with all of the JSON feed components.
}
\description{
Retrieve a JSON Feed
}
\examples{
get_feed("https://sixcolors.com/feed.json")

}
