% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_feed_info}
\alias{get_feed_info}
\title{Retrieve JSON feed meta data}
\usage{
get_feed_info(url)
}
\arguments{
\item{url}{the URL of a feed

Contains only the information about a feed.}
}
\value{
a list of the metadata about a feed. This does not contain post content.
}
\description{
Retrieve JSON feed meta data
}
\examples{
get_feed_info("https://sixcolors.com/feed.json")
}
