% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_feed_items}
\alias{get_feed_items}
\title{Retrieve JSON feed posts}
\usage{
get_feed_items(url)
}
\arguments{
\item{url}{the URL of a feed

A data.frame with all of the post items in a JSON feed.}
}
\value{
a data.frame of JSON feed posts
}
\description{
Retrieve JSON feed posts
}
\examples{
get_feed_items("https://sixcolors.com/feed.json")
}
